Separate server and browser code
While looking through some bundles built for browsers, I noticed that
this module brings in react-dom/server even though it isn't needed.
Minified, this appears to be adding ~13.5 KiB of bundle weight.

To avoid this, I split out the server and client code to separate
entrypoints, and then configured the browser field in package.json to
make the right module get used when imported.